1.7 login actions moved to CLI
main is pretty empty right now. There's a little bit of work left to do on separating the CLI from the core functionality.
