+++ src/CLI.java
@@ -7,7 +7,7 @@ public class CLI {
 
     static Scanner scanner = new Scanner(System.in);
 
-    public static void loginActionCLI() {
+    public static void userSelectionCLI() {
         boolean exit = false;
         while (!exit) {
             CLI.printOpeningMessage();
@@ -23,7 +23,7 @@ public class CLI {
                     CLI.deleteUserCLI();
                     break;
                 case 3:
-                    Main.login();
+                    login();
                     exit = true;
                     break;
                 case 4:
@@ -36,7 +36,66 @@ public class CLI {
         }
     }
 
-    public static void printWelcomeMessage () {
+    public static void login() {
+        System.out.println("Please enter your username: ");
+        String userInputName = scanner.nextLine();
+        boolean foundUser = false;
+
+        if (userInputName.isEmpty()) {
+            return;
+        }
+
+        for (User element : UserManager.userArrayList) {
+            if (element.userName.equals(userInputName)) {
+                foundUser = true;
+                break;
+            }
+        }
+
+        if (!foundUser) {
+            System.out.println("Sorry, I could not find " + userInputName);
+            return;
+        }
+
+        System.out.println("Please enter the password for " + userInputName);
+        String userInputPassword = scanner.nextLine();
+        boolean authenticated = UserManager.authenticateUserCLI(userInputName, userInputPassword);
+
+        if (!authenticated) {
+            System.out.println("Invalid password. Please try again.");
+            return;
+        }
+
+        taskSelector(userInputName);
+
+    }
+
+    public static void taskSelector(String userInputName){
+        System.out.println("Login successful. Welcome, " +userInputName);
+        while(true) {
+            CLI.CLIWelcomeMessage();
+            System.out.print("Enter a command (or 'exit' to save and exit): ");
+            String userInput = scanner.nextLine();
+            if (userInput.equals("GUI")) {
+                GUI.displayTasksGUI(userInputName);
+            } else if (userInput.equals("add")) {
+                CLI.addTaskCLI(userInputName);
+            } else if (userInput.equals("delete")) {
+                CLI.deleteTaskCLI(userInputName);
+            } else if (userInput.equals("complete")) {
+                CLI.completeTaskCLI(userInputName);
+            } else if (userInput.equals("list")) {
+                CLI.listTaskCLI(userInputName);
+            } else if (userInput.equals("exit")) {
+                TaskManager.exitAndSave();
+                break;
+            } else {
+                System.out.println("Sorry, I did not catch that!");
+            }
+        }
+    }
+
+    public static void CLIWelcomeMessage() {
         String filePathWelcome = "welcomeText.txt";
         try (BufferedReader reader = new BufferedReader(new FileReader(filePathWelcome))) {
             String line;
@@ -140,9 +199,7 @@ public class CLI {
         }
     }
 
-    //
-    // User management starts here!
-    //
+    /// User management starts here! ///
 
     public static void addUserCLI() {
         System.out.println("Please enter the username for the new user: ");
